defactor into coordinate utils
this allows typehiding but lets us ditch the fake return types.
